requirements.txt
pillow
requests
pytz
pyyaml
aiohttp
aiosqlite
